+++ src/CombatEventHandler.cpp
@@ -31,7 +31,7 @@ namespace TheLastBreath {
         }
 
         // Check if entering combat
-        if (a_event->newState.underlying() == 1) {  
+        if (a_event->newState.underlying() == 1) {
             std::lock_guard<std::mutex> lock(registrationMutex);
 
             auto formID = actor->GetFormID();
@@ -72,4 +72,4 @@ namespace TheLastBreath {
         return RE::BSEventNotifyControl::kContinue;
     }
 
-} 
+}